Refactored the website
Changed the Frontend to be Vue.js and the Backend to be Node.js. The server will now be MySQL and the UI Framework being used is Vuetify. Currently the website is boiler plate from Vueify with an added Registration Feature used to test the frontend, backend, and server that were implemented. The application is seperated between a server and a client side, with axios being used as the go-between. The server is currently using a RestAPI in order to send JSON to the front end, which will determine the information shown on the website. The only backend work being done currently is validating the input from the user during the registration. This is seperated by the policy that defines the rules for validation and the controller that executes the request.After entering the registration from the frontend, axios will send the registration data to the backend in order for the process explained previously to run.