Added extra inputs for registration and validation for registration inputs
Added a username and confirmation password in the registration page. Also added a way to show validation messages when input is not made properly. Currently not working completely since I am only doing the validation on the server side. Planning to change up the client and server files in order to have a easier time to share documents between them. This will allow for them to both do validation with a common validation middleware.