Fixed Validation On Registration Page
Now after clicking the register button, the user input is sent to the serve for validation. If an error is found, the error message is sent back to the client (json format with message name as label and the message as the body). This is then sent inputted to an errors data variable which will be used to determine hte message sent in the binded rules variable. The validation is done on button click, which means the error messages will be shown after clicking the button. Need to organize the inputs into their own structure and resetValidation with every button press.